fix: correctly initiate logging for service
make possible to use directly service without deploy
