+++ app/backend/services/main/__init__.py
@@ -1,16 +1,12 @@
 """
 Main KreoShine service
 """
-import logging
 import logging.config
-import sys
 
 from concurrent.futures import ThreadPoolExecutor
 
 import asyncio
 from aiohttp import web
-
-from settings import config
 
 logger = logging.getLogger('app')
 
@@ -36,28 +32,15 @@ async def on_app_stop(app) -> None:
     Args:
         app: instance of the application
     """
-    # todo: set log file path for 'service.log' as empty string
 
 
-def handle_exception(exc_type, exc_value, exc_traceback) -> None:
-    """ Handler for uncaught exceptions """
-    if issubclass(exc_type, KeyboardInterrupt):
-        sys.__excepthook__(exc_type, exc_value, exc_traceback)
-        return
-
-    logger.critical("Uncaught exception!", exc_info=(exc_type, exc_value, exc_traceback))
-
-
-def create_app(service_config: dict, log_config: dict) -> web.Application:
+def create_app(service_config: dict) -> web.Application:
     """
     Creates web application
 
     Args:
         service_config: configuration for service
     """
-    logging.config.dictConfig(config=log_config)
-    sys.excepthook = handle_exception
-
     app = web.Application(client_max_size=service_config['client_max_size'])
     app['service_config'] = service_config
     app.on_startup.append(on_app_start)
